+++ semi/WebContent/login/users_login.jsp
@@ -11,7 +11,7 @@
 <input class="block-item input-item" name ="user_id" placeholder="아이디 입력" required></div>
 <!-- PW 입력창 -->
 <div class="row">
-<input class="block-item input-item" name ="user_pw" placeholder="비밀번호 입력" required></div>
+<input class="block-item input-item" type="password" name ="user_pw" placeholder="비밀번호 입력" required></div>
 <!-- login 버튼-->
 <div class="row-left">
 	<input class="btn" type="submit" value="User Login"></div>
@@ -20,7 +20,7 @@
 <input type="checkbox">ID 저장하기</div>
 <!-- ID찾기 버튼 -->
 <div class="row-left">
-<a href="<%=request.getContextPath()%>/login/users_login_find.jsp">ID 찾기</a><br>
+<a href="<%=request.getContextPath()%>/login/users_find_id.jsp">ID 찾기</a><br>
 <a href="<%=request.getContextPath()%>/join/users_regist.jsp">Join us</a></div>
 <!-- error 전송될 경우 메세지 출력 -->
 <div>
