+++ semi/src/servlet/Users_Check_servlet.java
@@ -8,7 +8,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import beans.Users_Change_Info_Dao;
+import beans.Users_Info_Dao;
 
 //비밀번호 확인 서블릿
 @WebServlet(urlPatterns = "/info/users_check.do")
@@ -25,8 +25,8 @@ public class Users_Check_servlet extends HttpServlet {
 //			go= /info/users_change_pw.jsp or /info/users_change_info.jsp 
 			
 			//처리
-			Users_Change_Info_Dao UCIdao = new Users_Change_Info_Dao();
-			boolean result = UCIdao.login(user_id,user_pw);
+			Users_Info_Dao UCIdao = new Users_Info_Dao();
+			boolean result = UCIdao.users_login(user_id,user_pw);
 			
 			//이동
 			if(result) {//성공시
