+++ semi/src/servlet/Users_Regist_Servlet.java
@@ -8,7 +8,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import beans.Users_Regist_Dao;
+import beans.Users_Info_Dao;
 import beans.Users_Regist_Dto;
 
 @WebServlet(urlPatterns = "/join/users_regist.do")
@@ -28,8 +28,8 @@ public class Users_Regist_Servlet extends HttpServlet{
 			URdto.setUser_interest(req.getParameter("user_interest"));
 			URdto.setUser_birth(req.getParameter("user_birth"));
 			
-//			[2]dao를 불러와 메소드 사용, 가입 처리를 합니다
-			Users_Regist_Dao URdao = new Users_Regist_Dao();
+//			[2]dao를 불러와 users_regist메소드 사용, 가입 처리를 합니다
+			Users_Info_Dao URdao = new Users_Info_Dao();
 			URdao.users_regist(URdto);
 			
 //			[3]결과 페이지로 이동합니다
